requirements.txt
scipy
numpy
csc-pysparse
networkx
divisi2
python-recsys